refactor: remove redundant Id from TourPoint and align tests with inherited Entity Id

+++ src/Modules/Tours/Explorer.Tours.Infrastructure/Database/Repositories/TourPointDbRepository.cs
@@ -25,7 +25,7 @@ namespace Explorer.Tours.Infrastructure.Database.Repositories
 
         public TourPoint Get(int id)
         {
-            var p = _dbSet.Find(id);
+            var p = _dbSet.Find((long)id);
             if (p == null) throw new NotFoundException($"Point not found {id}");
             return p;
         }
